# frontend/public/ui_openapi.json
{
...
    "info": {
        "contact": {
            "name": "Notifiarr Discord",
            "url": "https://notifiarr.com/discord"
        },
        "description": "Updates the application configuration with new settings and optionally triggers a reload.",
        "license": {
            "name": "MIT",
            "url": "https://github.com/Notifiarr/notifiarr/blob/main/LICENSE"
        },
        "termsOfService": "https://notifiarr.com",
        "title": "Notifiarr Client GUI API Documentation",
        "version": "1.0"
    },
...
    "paths": {
...
        "/trigger/{trigger}": {
            "get": {
                "description": "Executes a specific trigger action.",
                "parameters": [
                    {
                        "description": "Trigger name to execute",
                        "in": "path",
                        "name": "trigger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional content for the trigger",
                        "in": "path",
                        "name": "content",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "trigger result"
                    },
                    "400": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "unknown trigger"
                    }
                },
                "summary": "Trigger action",
                "tags": [
                    "System"
                ]
            }
        },
        "/trigger/{trigger}/{content}": {
            "get": {
                "description": "Executes a specific trigger action.",
                "parameters": [
                    {
                        "description": "Trigger name to execute",
                        "in": "path",
                        "name": "trigger",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional content for the trigger",
                        "in": "path",
                        "name": "content",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "trigger result"
                    },
                    "400": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "unknown trigger"
                    }
                },
                "summary": "Trigger action",
                "tags": [
                    "System"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/ui"
        }
    ]
}